add services.py and ajax_jqery.js for sending POST requests from LinksForm  withoth reload index page
