+++ app/link_shortening/forms.py
@@ -4,11 +4,14 @@ from link_shortening.models import Links
 
 
 class LinksForm(forms.ModelForm):
-    long_link = forms.URLField(widget=forms.URLInput(attrs={"class": "link_field",
+    long_link = forms.URLField(widget=forms.URLInput(attrs={"class": "link_field", 
+                                                            "id": "link",
                                                             "placeholder": "URL"}))
-    time_to_delete = forms.IntegerField(min_value=1, max_value=365,
-                                        widget=forms.TextInput(attrs={"class": "timt_to_delete_fild",
-                                                                      "placeholder": "30"}))
+    
+    time_to_delete = forms.IntegerField(min_value=1, max_value=365, widget=forms.TextInput(
+                                                                    attrs={"class": "timt_to_delete_fild", 
+                                                                           "id": "delete",
+                                                                           "placeholder": "30"}))
 
     class Meta:
         model = Links
